offline_resume_builder/requirements.txt
Flask
pdfkit
python-docx
mammoth
requests
ollama
